+++ CHANGELOG.md
@@ -1,3 +1,58 @@
+## Version 2.0.0
+
+Many optimizations, bug fixes, and new features. With this version, the popular "calendar" showcase
+for range libraries is supported (see [./test/calendar.cpp](calendar.cpp) for more).
+
+### Features
+
+- Algorithms added:
+  - `chain()`
+  - `cycle()`
+  - `flatten()`
+  - `for_each()`
+  - `group_adjacent_by()`
+  - `in_groups_of_exactly()`
+  - `in_groups_of()`
+  - `null_sink()`
+  - `padded()`
+  - `tee()`
+  - `zip_longest()`
+
+### Bugfixes
+
+- Overload resolution for `as_input_range()` was very confusing. This has been cleaned up.
+- Some operations that were expected to be idempotent were not (#1).
+- Combinators now support non-default-constructible value types, unless the output explicitly
+  requires it (#7).
+- Fix `first()` after `sort()` (#8).
+- `sort()` can now be constructed with comparison predicates taking more than one constructor
+  argument.
+- `sort()`/`min()`/`max()` support non-default-constructible and non-copyable predicates.
+
+### Optimizations
+
+- When compilers support it, `__builtin_expect` is utilized in inner loops to aid code layout.
+- Idempotency is now only ensured when the input isn't already idempotent.
+- Fine-tuning of many internal loops to avoid unnecessary branches.
+- Where possible, the "empty base class" optimization is used. This mostly applies to combinators
+  that take a standard comparison predicate (`std::less`, `std::equal_to`, etc.).
+- `group_adjacent_by()` no longer requires internal temporary storage in a vector (#6).
+
+### Other
+
+- Grouping combinators now produce subranges instead of `std::array`s or `std::vector`s, eliminating
+  the need to allocate temporary internal storage.
+- Added a benchmark suite using Google Benchmark.
+- Cleanup of internal naming conventions.
+- `T::is_finite` is no longer required for input ranges (defaults to `false` if missing).
+- `min()` and `max()` can now operate with a custom comparison function.
+
+### Contributors
+
+- René Kijewski
+- Jules Ricou
+- Simon Ask Ulsnes
+
 ## Version 1.0.1
 
 Bug fixes and feature enhancements. Thanks for the valuable feedback from /r/cpp!